Código passado do App para componente Home
